Avalon

+++ Avalon/Avalon.Text/Infra.cs
@@ -390,7 +390,7 @@ public class Infra : Any
         return k;
     }
 
-    public virtual int LastIndex(Text text, Text other, InfraCompare compare)
+    public virtual long LastIndex(Text text, Text other, InfraCompare compare)
     {
         if (!this.ValidRange(text))
         {
@@ -404,12 +404,12 @@ public class Infra : Any
         Range textRange;
         textRange = text.Range;
 
-        int textIndex;
-        int textCount;
+        long textIndex;
+        long textCount;
         textIndex = textRange.Index;
         textCount = textRange.Count;
 
-        int otherCount;
+        long otherCount;
         otherCount = other.Range.Count;
 
         if (textCount < otherCount)
@@ -417,16 +417,16 @@ public class Infra : Any
             return -1;
         }
 
-        int k;
+        long k;
         k = -1;
 
-        int count;
+        long count;
         count = textCount - otherCount + 1;
-        int i;
+        long i;
         i = 0;
         while (k == -1 & i < count)
         {
-            int index;
+            long index;
             index = textIndex + count - 1 - i;
 
             textRange.Index = index;
